# Optimize Date plugin again (refresh timezone only once a minute)