--- a/xmobar.cabal
+++ b/xmobar.cabal
@@ -356,6 +356,6 @@ benchmark xmobarbench
   main-is: main.hs
   hs-source-dirs:
       bench
-  ghc-options: -O2
+  ghc-options: -funbox-strict-fields -Wall -fno-warn-unused-do-bind -O2
   build-depends: base, gauge, xmobar, mtl
   default-language: Haskell2010