--- a/xmobar.cabal
+++ b/xmobar.cabal
@@ -357,5 +357,5 @@ benchmark xmobarbench
   hs-source-dirs:
       bench
   ghc-options: -funbox-strict-fields -Wall -fno-warn-unused-do-bind -O2
-  build-depends: base, gauge, xmobar, mtl
+  build-depends: base, gauge, xmobar, mtl, time
   default-language: Haskell2010