@jaor 2020-11-13
Thanks for the very informative exchange and the attention to detail, guys.  In my opinion, this patch puts together the best of both worlds (same amount of code changes, almost same performance, much simpler user interface), so i'm merging and i think we can discard #502, unless @psibi feels strongly for it.